ref: Move SymCache/SourceLookup into their own file
